refactor(entity-patch-common): split index.ts into types, mappingUtils, and validation modules
- Extract all interface definitions into types.ts (PatchProperty, PatchSection, PatchDefinition, RelationPair, PatchConfig)
- Extract mapping utilities into mappingUtils.ts (isMappingTemplate, flattenMapping, buildRelationPairs, buildPatchConfigs)
- Extract zod schemas and validatePatchConfig into validation.ts
- Replace index.ts with re-exports — public API unchanged, no consumer changes needed
